Modified   include/Common.h Modified   include/Controls/Control.h Modified   include/Controls/ControlFactory.h Modified   include/Controls/LED.h Modified   include/Controls/Potentiometer.h Modified   include/DSP/Chorus.h Modified   include/DSP/Distortion.h Modified   include/DSP/Overdrive.h Modified   include/DSP/Reverb.h Modified   include/DSP/Sustain.h Modified   include/DSP/Tremolo.h Modified   include/Filters/BiquadFilter.h Modified   include/Filters/DelayFilter.h Modified   include/Filters/HighPassFilter.h Modified   include/Filters/LowPassFilter.h Modified   include/Filters/OscillatorFilter.h Modified   include/Math.h

+++ include/Filters/HighPassFilter.h
@@ -72,9 +72,9 @@ private:
 private:
 	uint32 m_SampleRate;
 	float m_DeltaTime;
-	double m_CutoffFrequency;
-	double m_Alpha;
-	double m_CapacitorVoltage;
+	float m_CutoffFrequency;
+	float m_Alpha;
+	float m_CapacitorVoltage;
 
 public:
 	static constexpr double MAX_CUTOFF_FREQUENCY = 14066.730469;
